refactor: split large shared and frontend modules
